updated routes to catch errors and use route query

+++ controllers/apiRoutes/project.js
@@ -3,32 +3,39 @@ var db = require('../../models');
 
 router.get('/', function (req, res) {
   //get project
-  db.Project.findOne({
-    where: {
-      id_project: req.params.id,
-    },
-  }).then(function (result) {
-    res.json(result);
-  });
+  if (req.query.id) {
+    db.Project.findOne({
+      where: {
+        id_project: req.query.id,
+      },
+    }).then(function (result) {
+      res.json(result);
+    });
+  } else {
+    res.status(401).json(err);
+  }
 });
 
 router.post('/', function (req, res) {
   //new project
-  var project = req.body;
-  db.Project.create({
-    id_project: project.id_project,
-    id_team: project.id_team,
-    project_name: project.project_name,
-    project_description: project.project_description,
-    start_date: project.start_date,
-    end_date: project.end_date,
-    enabled: project.enabled,
-    automation_order: project.automation_order,
-    created_by: project.created_by,
-  }).then((result) => {
-    console.log(result);
-    res.redirect('/');
-  });
+  if (req.body.project_name) {
+    var project = req.body;
+    db.Project.create({
+      id_team: project.id_team,
+      project_name: project.project_name,
+      project_description: project.project_description,
+      start_date: project.start_date,
+      end_date: project.end_date,
+      enabled: project.enabled,
+      automation_order: project.automation_order,
+      created_by: project.created_by,
+    }).then((result) => {
+      console.log(result);
+      res.redirect('/');
+    });
+  } else {
+    res.status(401).json(err);
+  }
 });
 
 router.put('/', function (req, res) {
@@ -36,7 +43,7 @@ router.put('/', function (req, res) {
   if (req.body.project_name) {
     db.Project.findOne({
       where: {
-        id_project: req.params.id,
+        id_project: req.query.id,
       },
     })
       .then(function (results) {
@@ -54,13 +61,15 @@ router.put('/', function (req, res) {
       .catch((err) => {
         res.status(401).json(err);
       });
+  } else {
+    res.status(401).json(err);
   }
   // update project description
 
   if (req.body.project_description) {
     db.Project.findOne({
       where: {
-        id_project: req.params.id,
+        id_project: req.query.id,
       },
     })
       .then(function (results) {
@@ -80,13 +89,15 @@ router.put('/', function (req, res) {
       .catch((err) => {
         res.status(401).json(err);
       });
+  } else {
+    res.status(401).json(err);
   }
   // update project start date
 
   if (req.body.start_date) {
     db.Project.findOne({
       where: {
-        id_project: req.params.id,
+        id_project: req.query.id,
       },
     })
       .then(function (results) {
@@ -104,13 +115,15 @@ router.put('/', function (req, res) {
       .catch((err) => {
         res.status(401).json(err);
       });
+  } else {
+    res.status(401).json(err);
   }
   // update project start date
 
   if (req.body.end_date) {
     db.Project.findOne({
       where: {
-        id_project: req.params.id,
+        id_project: req.query.id,
       },
     })
       .then(function (results) {
@@ -128,18 +141,26 @@ router.put('/', function (req, res) {
       .catch((err) => {
         res.status(401).json(err);
       });
+  } else {
+    res.status(401).json(err);
   }
 });
 
 router.delete('/', function (req, res) {
   //delete project
-  db.Project.destroy({
-    where: {
-      id_project: req.params.id,
-    },
-  }).then(function (result) {
-    res.json(result);
-  });
+  if (req.query.id) {
+    db.Project.destroy({
+      where: {
+        id_project: req.query.id,
+      },
+    })
+      .then(function (result) {
+        res.json(result);
+      })
+      .catch((err) => {
+        res.status(401).json(err);
+      });
+  }
 });
 
 module.exports = router;
